+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### vNext
 
--
+- Add onboarding feature flag and disable it on all envs
 
 ### [Developer Console] v0.10.0